@@ -46,16 +46,19 @@
         </repository>
     </distributionManagement>
     <properties>
-        <compiler-plugin.version>3.10.1</compiler-plugin.version>
+        <compiler-plugin.version>3.11.0</compiler-plugin.version>
         <maven.compiler.parameters>true</maven.compiler.parameters>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
-        <org.jooq.version>3.16.14</org.jooq.version>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
+        <org.jooq.version>3.18.7</org.jooq.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <surefire-plugin.version>3.0.0-M9</surefire-plugin.version>
-        <swagger.version>1.6.9</swagger.version>
-        <jooreo.version>0.2.15</jooreo.version>
-        <jakarta.json-bind.version>1.0.2</jakarta.json-bind.version>
+        <surefire-plugin.version>3.1.2</surefire-plugin.version>
+        <swagger.version>1.6.12</swagger.version>
+        <jooreo.version>1.0.0-rc.1</jooreo.version>
+        <jakarta.json-bind.version>3.0.0</jakarta.json-bind.version>
+        <jakarta.cdi.version>4.0.1</jakarta.cdi.version>
+        <jakarta.inject.version>2.0.1</jakarta.inject.version>
+        <jakarta.validation.version>3.0.2</jakarta.validation.version>
     </properties>
     <dependencies>
         <!-- jakarta deps -->
