@@ -27,7 +27,7 @@ package com.github.gpor0.jaffas.endpoints;
 
 import com.github.gpor0.jaffas.endpoints.model.SyncRolePermissions;
 
-import javax.ws.rs.*;
+import jakarta.ws.rs.*;
 
 /**
  * <p>No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
