@@ -1,12 +1,12 @@
 package com.github.gpor0.jaffas.exceptions.mappers;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.ext.Provider;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.ext.Provider;
 
 @Provider
 @ApplicationScoped
-public class ServiceExceptionMapper extends BaseExceptionMapper implements javax.ws.rs.ext.ExceptionMapper<Exception> {
+public class ServiceExceptionMapper extends BaseExceptionMapper implements jakarta.ws.rs.ext.ExceptionMapper<Exception> {
 
     @Override
     public Response toResponse(Exception e) {
