@@ -24,7 +24,7 @@
 
 package com.github.gpor0.jaffas.endpoints.model;
 
-import javax.json.bind.annotation.JsonbProperty;
+import jakarta.json.bind.annotation.JsonbProperty;
 import java.util.ArrayList;
 import java.util.List;
 
