@@ -1,6 +1,6 @@
 package com.github.gpor0.jaffas.endpoints.model;
 
-import javax.json.bind.annotation.JsonbProperty;
+import jakarta.json.bind.annotation.JsonbProperty;
 
 public class Token {
 
